move tests from actions to bundle
