+++ rsconnect/actions.py
@@ -109,6 +109,7 @@ def which_python(python, env=os.environ):
     * RETICULATE_PYTHON defined in the environment
     * the python binary running this script
     """
+    warn("This method has been moved and will be deprecated.", DeprecationWarning, stacklevel=2)
     if python:
         if not (exists(python) and os.access(python, os.X_OK)):
             raise RSConnectException('The file, "%s", does not exist or is not executable.' % python)
