+++ tests/test_actions.py
@@ -1,15 +1,13 @@
 import os
-import shutil
 import subprocess
 import sys
-import tempfile
 
 try:
     import typing
 except ImportError:
     typing = None
 
-from os.path import basename, dirname, join
+from os.path import basename, join
 from unittest import TestCase
 
 import pytest
@@ -19,11 +17,8 @@ import rsconnect.actions
 from rsconnect import api
 
 from rsconnect.actions import (
-    _default_title,
-    _default_title_from_manifest,
     _make_deployment_name,
     _to_server_check_list,
-    _validate_title,
     _verify_server,
     are_apis_supported_on_server,
     check_server_capabilities,
@@ -35,31 +30,19 @@ from rsconnect.actions import (
     deploy_bokeh_app,
     gather_basic_deployment_info_for_api,
     get_python_env_info,
-    inspect_environment,
     is_conda_supported_on_server,
-    validate_entry_point,
-    validate_extra_files,
-    which_python,
 )
 from rsconnect.api import RSConnectServer
 from rsconnect.environment import MakeEnvironment
 from rsconnect.exception import RSConnectException
 
-from .utils import get_manifest_path, get_api_path, get_dir
+from .utils import get_api_path, get_dir
 
 
 class TestActions(TestCase):
     @staticmethod
     def optional_target(default):
         return os.environ.get("CONNECT_DEPLOY_TARGET", default)
-
-    def test_which_python(self):
-        with self.assertRaises(RSConnectException):
-            which_python("fake.file")
-
-        self.assertEqual(which_python(sys.executable), sys.executable)
-        self.assertEqual(which_python(None), sys.executable)
-        self.assertEqual(which_python(None, {"RETICULATE_PYTHON": "fake-python"}), "fake-python")
 
     def test_verify_server(self):
         with self.assertRaises(RSConnectException):
@@ -82,7 +65,7 @@ class TestActions(TestCase):
         no_api_support = {"python": {"api_enabled": False}}
         api_support = {"python": {"api_enabled": True}}
 
-        with self.assertRaises(api.RSConnectException) as context:
+        with self.assertRaises(RSConnectException) as context:
             check_server_capabilities(None, (are_apis_supported_on_server,), lambda x: no_api_support)
         self.assertEqual(
             str(context.exception),
@@ -95,14 +78,14 @@ class TestActions(TestCase):
         conda_not_supported = {"conda": {"supported": False}}
         conda_supported = {"conda": {"supported": True}}
 
-        with self.assertRaises(api.RSConnectException) as context:
+        with self.assertRaises(RSConnectException) as context:
             check_server_capabilities(None, (is_conda_supported_on_server,), lambda x: no_conda)
         self.assertEqual(
             str(context.exception),
             "Conda is not supported on the target server.  " + "Try deploying without requesting Conda.",
         )
 
-        with self.assertRaises(api.RSConnectException) as context:
+        with self.assertRaises(RSConnectException) as context:
             check_server_capabilities(None, (is_conda_supported_on_server,), lambda x: conda_not_supported)
         self.assertEqual(
             str(context.exception),
@@ -120,50 +103,19 @@ class TestActions(TestCase):
             """A docstring."""
             return False
 
-        with self.assertRaises(api.RSConnectException) as context:
+        with self.assertRaises(RSConnectException) as context:
             check_server_capabilities(None, (fake_cap,), lambda x: None)
         self.assertEqual(
             str(context.exception),
             "The server does not satisfy the fake_cap capability check.",
         )
 
-        with self.assertRaises(api.RSConnectException) as context:
+        with self.assertRaises(RSConnectException) as context:
             check_server_capabilities(None, (fake_cap_with_doc,), lambda x: None)
         self.assertEqual(
             str(context.exception),
             "The server does not satisfy the fake_cap_with_doc capability check.",
         )
-
-    def test_validate_title(self):
-        with self.assertRaises(RSConnectException):
-            _validate_title("12")
-
-        with self.assertRaises(RSConnectException):
-            _validate_title("1" * 1025)
-
-        _validate_title("123")
-        _validate_title("1" * 1024)
-
-    def test_validate_entry_point(self):
-        directory = tempfile.mkdtemp()
-
-        try:
-            self.assertEqual(validate_entry_point(None, directory), "app")
-            self.assertEqual(validate_entry_point("app", directory), "app")
-            self.assertEqual(validate_entry_point("app:app", directory), "app:app")
-
-            with self.assertRaises(RSConnectException):
-                validate_entry_point("x:y:z", directory)
-
-                with open(join(directory, "onlysource.py"), "w") as f:
-                    f.close()
-                    self.assertEqual(validate_entry_point(None, directory), "onlysource")
-
-                    with open(join(directory, "main.py"), "w") as f:
-                        f.close()
-                        self.assertEqual(validate_entry_point(None, directory), "main")
-        finally:
-            shutil.rmtree(directory)
 
     def test_make_deployment_name(self):
         self.assertEqual(_make_deployment_name(None, "title", False), "title")
@@ -176,44 +128,6 @@ class TestActions(TestCase):
         self.assertEqual(_make_deployment_name(None, "M\ry\n \tT\u2103itle", False), "my_title")
         self.assertEqual(_make_deployment_name(None, "\r\n\t\u2103", False), "___")
         self.assertEqual(_make_deployment_name(None, "\r\n\tR\u2103", False), "__r")
-
-    def test_default_title(self):
-        self.assertEqual(_default_title("testing.txt"), "testing")
-        self.assertEqual(_default_title("this.is.a.test.ext"), "this.is.a.test")
-        self.assertEqual(_default_title("1.ext"), "001")
-        self.assertEqual(_default_title("%s.ext" % ("n" * 2048)), "n" * 1024)
-
-    def test_default_title_from_manifest(self):
-        self.assertEqual(_default_title_from_manifest({}, "dir/to/manifest.json"), "0to")
-        # noinspection SpellCheckingInspection
-        m = {"metadata": {"entrypoint": "point"}}
-        self.assertEqual(_default_title_from_manifest(m, "dir/to/manifest.json"), "point")
-        m = {"metadata": {"primary_rmd": "file.Rmd"}}
-        self.assertEqual(_default_title_from_manifest(m, "dir/to/manifest.json"), "file")
-        m = {"metadata": {"primary_html": "page.html"}}
-        self.assertEqual(_default_title_from_manifest(m, "dir/to/manifest.json"), "page")
-        m = {"metadata": {"primary_wat?": "my-cool-thing.wat"}}
-        self.assertEqual(_default_title_from_manifest(m, "dir/to/manifest.json"), "0to")
-        # noinspection SpellCheckingInspection
-        m = {"metadata": {"entrypoint": "module:object"}}
-        self.assertEqual(_default_title_from_manifest(m, "dir/to/manifest.json"), "0to")
-
-    def test_validate_extra_files(self):
-        # noinspection SpellCheckingInspection
-        directory = dirname(get_manifest_path("shinyapp"))
-
-        with self.assertRaises(RSConnectException):
-            validate_extra_files(directory, ["../other_dir/file.txt"])
-
-        with self.assertRaises(RSConnectException):
-            validate_extra_files(directory, ["not_a_file.txt"])
-
-        self.assertEqual(validate_extra_files(directory, None), [])
-        self.assertEqual(validate_extra_files(directory, []), [])
-        self.assertEqual(
-            validate_extra_files(directory, [join(directory, "index.htm")]),
-            ["index.htm"],
-        )
 
     def test_deploy_python_api_validates(self):
         directory = get_api_path("flask")
@@ -256,11 +170,6 @@ class TestActions(TestCase):
             create_api_deployment_bundle(directory, [], [], "bogus:bogus:bogus", None, None, None, None)
         with self.assertRaises(RSConnectException):
             create_api_deployment_bundle(directory, ["bogus"], [], "app:app", MakeEnvironment(), None, True, None)
-
-    def test_inspect_environment(self):
-        environment = inspect_environment(sys.executable, get_dir("pip1"))
-        assert environment is not None
-        assert environment.python != ""
 
 
 @pytest.mark.parametrize(
@@ -355,7 +264,7 @@ def test_get_python_env_info(
     monkeypatch.setattr(rsconnect.actions, "which_python", fake_which_python)
 
     if expected_environment.error is not None:
-        with pytest.raises(api.RSConnectException):
+        with pytest.raises(RSConnectException):
             _, _ = get_python_env_info(file_name, python, conda_mode=conda_mode, force_generate=force_generate)
     else:
         python, environment = get_python_env_info(
